requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
crewai>=0.11.0
langchain-openai>=0.0.2
python-dotenv>=1.0.0
redis>=4.0.0
rq>=1.15.0
beautifulsoup4>=4.9.3
requests>=2.25.1 